Updates to register modal
