+++ admin_app/src/utils/api.ts
@@ -12,6 +12,7 @@ const api = axios.create({
 });
 
 import { AxiosResponse, AxiosError } from "axios";
+import { is } from "date-fns/locale";
 
 api.interceptors.response.use(
   (response: AxiosResponse) => response,
@@ -66,19 +67,19 @@ const getLoginToken = async (username: string, password: string) => {
     throw new Error("Error fetching login token");
   }
 };
+
 const registerUser = async (username: string, password: string) => {
   try {
     const response = await api.post(
-      "/user",
-      { username: username, password: password },
+      "/user/register-first-user",
+      { username, password, is_admin: true },
       {
         headers: { "Content-Type": "application/json" },
       },
     );
     return response.data;
   } catch (error) {
-    console.log(error);
-    throw new Error("Error fetching login token");
+    console.error(error);
   }
 };
 const getGoogleLoginToken = async (idToken: {
